% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_funcs.R
\name{add_constraint}
\alias{add_constraint}
\title{Add a constraint to a constraint data frame object and open for editing.}
\usage{
add_constraint(dt, constraint, type = "chr", default = "")
}
\arguments{
\item{dt}{Data frame object}

\item{constraint}{Name of the new constraint to be added}

\item{type}{The type of constraint it is. "int" for integer, "bool" for boolean}

\item{default}{The default value to give the constraint}
}
\value{
Updated constraint data frame object
}
\description{
Add a constraint to a constraint data frame object and open for editing.
}
