% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_param_names.R
\name{combine_param_names}
\alias{combine_param_names}
\title{Making a list of possible parameters.}
\usage{
combine_param_names(func1, func2)
}
\arguments{
\item{func1}{A function.}

\item{func2}{A function.}
}
\value{
A vector of possible parameter names
}
\description{
Making a list of possible parameters.
}
\examples{
possible_parameternames<- combine_param_names(pickmdl::x13_both, RJDemetra::x13_spec)
}
