% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_funcs.R
\name{edit_constraints}
\alias{edit_constraints}
\title{Edit a constraint data frame object}
\usage{
edit_constraints(dt)
}
\arguments{
\item{dt}{Data frame object}
}
\value{
Updated data frame object
}
\description{
Edit a constraint data frame object
}
