% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ledige}
\alias{ledige}
\title{Ledige Stillinger}
\format{
A time series object with 60 rows and 51 variables:
\describe{
  \item{..}{noe }
}
}
\usage{
ledige
}
\description{
A time series object containing various Norwegian vacant positions from K1 2010 to K1 2025. Hentet fra statistikkbanken 07.04.25
}
\keyword{datasets}
