% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_plot_fig.R
\name{make_plot_fig}
\alias{make_plot_fig}
\title{Interactive plots of seasonally adjusted time series.}
\usage{
make_plot_fig(
  models_in,
  series_now,
  plot_start = NULL,
  linearized = F,
  cal_adjust = F,
  ma_filter = F
)
}
\arguments{
\item{models_in}{A list of output objects from the x13()-function.}

\item{series_now}{Names of series to be plotted.}

\item{plot_start}{Start date of time axis in plot.}

\item{linearized}{Linearized times series to be shown in plots? Default is FALSE.}

\item{cal_adjust}{Calender adjusted times series to be shown in plots? Default is FALSE.}

\item{ma_filter}{1x3 MA-filter to be shown in plots? Default is FALSE.}
}
\value{
List of plotly objects.
}
\description{
A function that creates a list of interactive plots of time series seasonally adjusted with RJDemetra.
Interactive plots are made with Plotly's R graphing library.
Interactive plots show unadjusted series, seasonally adjusted series and trend.
}
