% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_funcs.R
\name{read_yaml_constraints}
\alias{read_yaml_constraints}
\title{Read in yaml file with constraints information}
\usage{
read_yaml_constraints(path)
}
\arguments{
\item{path}{String variable for the path and name of yaml contraints file.}
}
\value{
A data frame object with the constrainst data as a data.frame object
}
\description{
Read in yaml file with constraints information
}
