% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_quality_report.R
\name{sa_quality_report}
\alias{sa_quality_report}
\title{Quality Report for Seasonal Adjustment with RJDemetra.}
\usage{
sa_quality_report(
  models_in,
  report_file,
  title = NULL,
  author = NULL,
  group_series = NULL,
  plots_included = TRUE,
  plot_start = NULL,
  outlier_choiche = 1,
  spec_file = NULL,
  linearized = FALSE,
  cal_adjust = FALSE,
  ma_filter = FALSE,
  n_digits = 2
)
}
\arguments{
\item{models_in}{List of output objects from x13_pickmdl()-function. \cr
Must be named, i.e. names(models_in) can not be empty.}

\item{report_file}{Name of output file, including path.}

\item{title}{Title of the report. Default is NULL.}

\item{author}{Author of the report. Default is NULL.}

\item{group_series}{List of vectors with model names that defines structuring of tables and plots. \cr
If NULL, all models are included in a single table. See examples. Default is NULL. \cr}

\item{plots_included}{If TRUE, interactive plots are included in the report. Default is TRUE}

\item{plot_start}{Start date of time axis in plot. If NULL, the whole time series is plotted. Default is NULL.}

\item{outlier_choiche}{How outliers are counted \cr
\itemize{
 \item 1: All outliers are counted.
 \item 2: When identifcation_end = TRUE and identify_outliers = TRUE (default), only outliers after identification end are counted, i.e. only after date of ARIMA model choice.
 When identification_end = TRUE and identify_outliers = FALSE, all outliers that are not pre-specified are counted.
 When identification_end = FALSE, no outliers are counted.
 \item 3: When corona = TRUE, only outliers outside corona period are counted. When corona = FALSE, all outliers are counted.
 \item 4: All outliers that are not pre-specified are counted.}
 Default is 1. \cr}

\item{spec_file}{When outlier_choiche is 3 or 4, data frame with specifications must be given as input. Default is NULL.}

\item{linearized}{Linearized series to be shown in plots? Default is FALSE. \cr
"decomposition.b1" must be userdefined in x13_pickmdl() for linearized to be calculated.}

\item{cal_adjust}{Calender adjusted times series to be shown in plots? Default is FALSE. \cr
"decomposition.a8" and "decomposition.b1" must be userdefined in x13_pickmdl() for cal_adjust to be calculated.}

\item{ma_filter}{1x3 symmetric MA-filter to be shown in plots? Default is FALSE.}

\item{n_digits}{number of printed digits. Default is 2.}
}
\value{
A html-document with quality report created at path given in parameter report_file.
}
\description{
Function for creating a quality report for seasonal adjustment with RJdemetra.
}
\examples{

time_series <- sadashboard::vhi
spec_now <- RJDemetra::x13_spec("RSA5c")

my_models <- list()

# Friedman test and independence of residuals must be user defined as output from x13_pickmdl().
# Warning if not user defined as part of output.

for(i in 1:ncol(time_series)){
  time_series_now <- time_series[,i]
  my_models[[i]] <- pickmdl::x13_pickmdl(time_series_now,spec_now,corona=FALSE,
                       pickmdl_method="first_tryautomdl",
                       userdefined = c("decomposition.a8","decomposition.b1",
                                "diagnostics.seas-sa-friedman","residuals.independence.value"))
}

names(my_models) <- colnames(time_series)

# Grouping of models
group_key <- c(paste0("47.",1:7),"47.9")
groups_now <- lapply(group_key,
                 function(x){colnames(time_series)[which(grepl(x,colnames(time_series)))]})
names(groups_now) <- group_key

path_now <- getwd()

sa_quality_report(models_in= my_models,report_file = paste0(path_now,"/my_report.html"),
                     group_series = groups_now,
                     title = "Eksempelrapport", author = "SSB")
}
