% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/td_p.R
\name{td_p}
\alias{td_p}
\title{Test of trading day variables in RegARIMA}
\usage{
td_p(model_now)
}
\arguments{
\item{model_now}{output from x13() or x13_pickmdl().}
}
\value{
P-value of the F-test.
}
\description{
F-test of trading days variables in RegARIMA pre-processing with RJDemetra. \cr
\eqn{H_{0}}: There are no trading days effects. \cr
Trading days variables as defined by JDemetra or as by user with pickmdl::konstruksjon().
}
\examples{
time_series <- sadashboarddemo::vhi[,"47"]
my_model <- RJDemetra::x13(time_series,spec="RSA5c")

td_p(my_model)

}
