% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vhi}
\alias{vhi}
\title{Norwegian Retail Indexes}
\format{
R time series object with 35 columns and 302 rows.
Columns contains various Norwegian unadjusted value indexes in NACE group 47, retail trade, except of motor vehicles and motorcycles.
Indices are produced monthly by Statistics Norway.
}
\source{
From Statbank Norway. (April 11, 2025), https://www.ssb.no/en/statbank/table/07129
}
\usage{
vhi
}
\description{
A time series object with unadjusted monthly Norwegian Retail Indexes from January 2000 to February 2025.
}
\keyword{datasets}
