% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit_funcs.R
\name{write_yaml_constraints}
\alias{write_yaml_constraints}
\title{Write constraints data.frame object to disk as a yaml file}
\usage{
write_yaml_constraints(df, path)
}
\arguments{
\item{df}{Data frame object}

\item{path}{Path and file name to save the object as}
}
\description{
Write constraints data.frame object to disk as a yaml file
}
