% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ledige}
\alias{ledige}
\title{Job Vacancies}
\format{
R time series object with 21 columns and 60 rows.
Columns contains various indexes for Norwegian job vacancy rates in various major industries (SIC2007).
Indices are produced quarterly by Statistics Norway.
}
\source{
From Statbank Norway. (April 11, 2025), https://www.ssb.no/en/statbank/table/08771/
}
\usage{
ledige
}
\description{
A time series object with unadjusted quarterly Norwegian job vacancy rates. Indexes from first quarter 2010 to fourth quarter 2024.
}
\keyword{datasets}
