% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quality_df.R
\name{make_quality_df}
\alias{make_quality_df}
\title{Quality tables for Seasonal Adjustment with RJDemetra.}
\usage{
make_quality_df(models_in, n_digits = 2, outlier_choiche = 1, spec_file = NULL)
}
\arguments{
\item{models_in}{list of relevant models. List of output objects from x13_pickmdl()-function. \cr
See examples for details.}

\item{n_digits}{number of printed digits. Default is 2.}

\item{outlier_choiche}{how to count outliers. Default is 1.}

\item{spec_file}{data frame with specifications. Only needed when outlier_choiche is set to 3 or 4. This is the data frame with specifications as used in x13_text_frame().}
}
\value{
A list of data frames.
}
\description{
Function creates two data frames with selected quality indicators for seasonal adjustment with RJDemetra at SSB.
First data frame contains quality indicators for sesonality and residual seasonality.
Second data frame contains quality indicators for pre-processing and RegARIMA-model.
}
\examples{

time_series <- sadashboard::vhi
spec_now <- RJDemetra::x13_spec("RSA5c")

my_models <- list()

# Friedman test and independence of residuals must be user defined as output from x13_pickmdl().
# Warning if not user defined as part of output.

for(i in 1:ncol(time_series)){
  time_series_now <- time_series[,i]
  my_models[[i]] <- pickmdl::x13_pickmdl(time_series_now,spec_now,corona=FALSE,
                       pickmdl_method="first_tryautomdl",
                       userdefined = c("decomposition.a8","decomposition.b1",
                                "diagnostics.seas-sa-friedman","residuals.independence.value"))
}

names(my_models) <- colnames(time_series)

my_quality <- make_quality_df(my_models)

seasonal_indicators <- my_quality[[1]]
regarima_indicators <- my_quality[[2]]
}
