% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ts_df.R
\name{make_ts_df}
\alias{make_ts_df}
\title{Data frame with seasonally adjusted time series}
\usage{
make_ts_df(model_now, linearized = F, cal_adjust = F, ma_filter = F)
}
\arguments{
\item{model_now}{Output from x13() or x13_pickmdl().}

\item{linearized}{Linearized times series (B1) to be included in data frame? Default is FALSE.
Only available if decomposition.b1 is included in userdefined in x13().}

\item{cal_adjust}{Calender adjusted times series to be included in data frame? Default is FALSE.
Only available if decomposition.a8 and decomposition.b1 are included in userdefined in x13().}

\item{ma_filter}{1x3 symmetric MA-filter to be included in data frame? Default is FALSE.}
}
\value{
A data frame with date and selected time series.
}
\description{
A function that creates a data frame with seasonally adjusted series from x13().
Columns in the data frame are date, unadjusted series, seasonally adjusted series and trend.
Optionally, linearized series, calendar adjusted series and 1x3 assymmetric MA-filter may be included.
}
\examples{

time_series <- sadashboard::vhi[,"47.3"]
my_model <- RJDemetra::x13(time_series,spec="RSA5c")
make_ts_df(my_model)

# Include linearized series.
my_model <- RJDemetra::x13(time_series,spec="RSA5c",
             userdefined=c("decomposition.b1"))
make_ts_df(my_model,cal_adjust=TRUE)

# Include calendar adjusted series.
my_model <- RJDemetra::x13(time_series,spec="RSA5c",
             userdefined=c("decomposition.a8","decomposition.b1"))
make_ts_df(my_model,cal_adjust=TRUE)

# Include assymetric 1x3 MA-filter
make_ts_df(my_model,ma_filter=TRUE)

}
