% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_workspace_wrap.R
\name{save_workspace_wrap}
\alias{save_workspace_wrap}
\title{Save RJDemetra output as Workspace}
\usage{
save_workspace_wrap(models_in, wk_names, wk_file)
}
\arguments{
\item{models_in}{A list of output objects from the x13()-function. One list for each multi-document to be created in workspace.}

\item{wk_names}{Vector with the names of the multi-documents to be created in workspace}

\item{wk_file}{File name (.xml), including path.}
}
\description{
A wrapper function that creates and saves a workspace with the selected RJDemetra output.
The output is saved in a .xml-file that can be opened in JDemetra+, the graphical user interface.
}
\examples{

my_sa_models <- list()

vhi_ts <- sadashboard::vhi

spec_file_vhi <- make_paramfile(inndat = vhi_ts,spec= "RSA3")
my_sa_models[[1]] <- x13_text_frame(spec_file_vhi,series= "vhi_ts")

ledige_ts <- sadashboard::ledige
spec_file_ledige <- make_paramfile(inndat = ledige_ts,spec= "RSA3")
my_sa_models[[2]] <- x13_text_frame(spec_file_ledige,series= "ledige_ts")

save_workspace_wrap(models_in = my_sa_models,wk_names = c("vhi","ledige"),wk_file=paste0(getwd(),"/wk_example.xml"))
}
