% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vhi}
\alias{vhi}
\title{Norwegian Retail Indexes}
\format{
A time series object with some rows and some variables:
\describe{
  \item{47}{quality of the cut (Fair, Good, Very Good, Premium, Ideal)}
}
}
\usage{
vhi
}
\description{
A time series object containing various Norwegian retail indexes from January 2000 to January 2025. Hentet fra statistikkbanken 07.04.25
}
\keyword{datasets}
